@@ -5,9 +5,9 @@
         "laravel",
         "json-rpc"
     ],
-    "license": "MIT",
+    "license": "BSD-2-Clause",
     "require": {
-        "php": ">=5.5.0",
+        "php": ">=7.0",
         "illuminate/contracts": "^5.1",
         "illuminate/support": "^5.1",
         "illuminate/http": "^5.1",
@@ -15,7 +15,7 @@
         "psr/log": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "phpunit/phpunit": "~6.0"
     },
     "autoload": {
         "psr-4": {
