@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 use Upgate\LaravelJsonRpc\Server\Batch;
 use Upgate\LaravelJsonRpc\Contract\RequestFactoryInterface;
@@ -6,7 +7,7 @@ use Upgate\LaravelJsonRpc\Contract\RequestInterface;
 use Upgate\LaravelJsonRpc\Contract\RequestExecutorInterface;
 use Illuminate\Contracts\Support\Arrayable;
 
-class BatchTest extends PHPUnit_Framework_TestCase
+class BatchTest extends \PHPUnit\Framework\TestCase
 {
 
     public function testBatch()
@@ -27,14 +28,27 @@ class BatchTest extends PHPUnit_Framework_TestCase
         $requestExecutor->method('execute')->willReturnCallback(
             function ($requestMock) {
                 $response = $this->getMockBuilder(Arrayable::class)->getMock();
-                $response->method('toArray')->willReturn(['value' => $requestMock->mockValue]);
+                $response->method('toArray')->willReturn($requestMock->mockValue->params);
 
                 return $response;
             }
         );
         /** @var RequestExecutorInterface $requestExecutor */
 
-        $batch = new Batch(['foo', 'bar'], $requestFactory);
+        $foo = (object)[
+            'jsonrpc' => '2.0',
+            'method'  => 'foo',
+            'params'  => ['value' => 'foo'],
+            'id'      => 'foo',
+        ];
+        $bar = (object)[
+            'jsonrpc' => '2.0',
+            'method'  => 'bar',
+            'params'  => ['value' => 'bar'],
+            'id'      => 'bar',
+        ];
+
+        $batch = new Batch([$foo, $bar], $requestFactory);
         $response = $batch->executeWith($requestExecutor);
         $responseJson = $response->toJson();
         $responseData = json_decode($responseJson);
