@@ -1,9 +1,10 @@
 <?php
+declare(strict_types=1);
 
 use Upgate\LaravelJsonRpc\Server\MiddlewareAliasRegistry;
 use Upgate\LaravelJsonRpc\Contract\MiddlewaresConfigurationInterface;
 
-class MiddlewareAliasRegistryTest extends PHPUnit_Framework_TestCase
+class MiddlewareAliasRegistryTest extends \PHPUnit\Framework\TestCase
 {
 
     public function testMiddlewareAliasResolver()
@@ -19,12 +20,14 @@ class MiddlewareAliasRegistryTest extends PHPUnit_Framework_TestCase
         );
 
         $this->assertEquals('FooMiddleware', $middlewareAliasRegistry->findNameByAlias('foo'));
-        $this->assertNull($middlewareAliasRegistry->findNameByAlias('does_not_exist'));
 
         $this->assertEquals(
             ['FooMiddleware', 'BarMiddleware', 'BazMiddleware'],
             $middlewareAliasRegistry->resolveAliases(['foo', 'BarMiddleware', 'baz'])
         );
+
+        $this->expectException(\InvalidArgumentException::class);
+        $middlewareAliasRegistry->findNameByAlias('does_not_exist');
     }
 
 }
