@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Upgate\LaravelJsonRpc\Contract;
 
@@ -9,28 +10,29 @@ interface MiddlewaresConfigurationInterface
      * @param MiddlewareAliasRegistryInterface|null $aliases
      * @return $this
      */
-    public function setMiddlewareAliases(MiddlewareAliasRegistryInterface $aliases = null);
+    public function setMiddlewareAliases(MiddlewareAliasRegistryInterface $aliases = null
+    ): MiddlewaresConfigurationInterface;
 
     /**
      * @return string[]
      */
-    public function getMiddlewares();
+    public function getMiddlewares(): array;
 
     /**
      * @return bool
      */
-    public function isEmpty();
+    public function isEmpty(): bool;
 
     /**
      * @param array $middlewares
      * @return $this
      */
-    public function setMiddlewares(array $middlewares = []);
+    public function setMiddlewares(array $middlewares = []): MiddlewaresConfigurationInterface;
 
     /**
      * @param string $middleware
      * @return $this
      */
-    public function addMiddleware($middleware);
+    public function addMiddleware(string $middleware): MiddlewaresConfigurationInterface;
 
 }
