@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Upgate\LaravelJsonRpc\Contract;
 
@@ -6,42 +7,48 @@ interface RouteRegistryInterface
 {
 
     /**
-     * @param array|string $middleware
+     * @param string $middleware
      * @return $this
      */
-    public function addMiddleware($middleware);
+    public function addMiddleware(string $middleware): RouteRegistryInterface;
+
+    /**
+     * @param array $middlewares
+     * @return $this
+     */
+    public function addMiddlewares(array $middlewares): RouteRegistryInterface;
 
     /**
      * @param MiddlewareAliasRegistryInterface|null $aliases
      * @return $this
      */
-    public function setMiddlewareAliases(MiddlewareAliasRegistryInterface $aliases = null);
+    public function setMiddlewareAliases(MiddlewareAliasRegistryInterface $aliases = null): RouteRegistryInterface;
 
     /**
      * @param string $method
      * @param string $binding
      * @return $this
      */
-    public function bind($method, $binding);
+    public function bind(string $method, string $binding): RouteRegistryInterface;
 
     /**
      * @param string $namespace
      * @param string $controller
      * @return $this
      */
-    public function bindController($namespace, $controller);
+    public function bindController(string $namespace, string $controller): RouteRegistryInterface;
 
     /**
      * @param callable|array|string|null $middlewaresConfigurator
      * @param callable $routesConfigurator
      * @return $this
      */
-    public function group($middlewaresConfigurator = null, callable $routesConfigurator);
+    public function group($middlewaresConfigurator = null, callable $routesConfigurator): RouteRegistryInterface;
 
     /**
      * @param string $method
      * @return RouteInterface
      */
-    public function resolve($method);
+    public function resolve(string $method): RouteInterface;
 
 }
