@@ -1,18 +1,21 @@
 <?php
+declare(strict_types=1);
 
 namespace Upgate\LaravelJsonRpc\Exception;
+
+use Upgate\LaravelJsonRpc\Server\ErrorCode;
 
 final class ParseErrorException extends JsonRpcException
 {
 
-    protected function getDefaultMessage()
+    protected function getDefaultMessage(): string
     {
         return 'Parse error';
     }
 
-    protected function getDefaultCode()
+    protected function getDefaultCode(): int
     {
-        return -32700;
+        return ErrorCode::PARSE_ERROR;
     }
 
 }
