@@ -1,27 +1,27 @@
 <?php
+declare(strict_types=1);
 
 namespace Upgate\LaravelJsonRpc\Exception;
+
+use Upgate\LaravelJsonRpc\Server\ErrorCode;
 
 class ServerErrorException extends JsonRpcException
 {
 
-    protected function getDefaultMessage()
+    public function __construct(string $message = "", int $code = 0, \Exception $previous = null)
+    {
+        $code = $code ?: $this->getDefaultCode();
+        parent::__construct($message, $code, $previous);
+    }
+
+    protected function getDefaultMessage(): string
     {
         return 'Internal error';
     }
 
-    public function __construct($message = "", $code = 0, \Exception $previous = null)
+    protected function getDefaultCode(): int
     {
-        $code = $code ?: $this->getDefaultCode();
-        if ($code > -32000 || $code < -32099) {
-            throw new \InvalidArgumentException("Code out of range");
-        }
-        parent::__construct($message, $code, $previous);
-    }
-
-    protected function getDefaultCode()
-    {
-        return -32000;
+        return ErrorCode::INTERNAL_ERROR;
     }
 
 }
