@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Upgate\LaravelJsonRpc\Server;
 
@@ -9,9 +10,6 @@ class MiddlewareAliasRegistry implements MiddlewareAliasRegistryInterface
 
     private $registry = [];
 
-    /**
-     * @param array|null $aliases
-     */
     public function __construct(array $aliases = null)
     {
         if ($aliases) {
@@ -19,24 +17,14 @@ class MiddlewareAliasRegistry implements MiddlewareAliasRegistryInterface
         }
     }
 
-    /**
-     * @param string $alias
-     * @param string $name
-     * @return $this
-     */
-    public function registerAlias($alias, $name)
+    public function registerAlias(string $alias, string $name): MiddlewareAliasRegistryInterface
     {
         $this->registry[$alias] = (string)$name;
 
         return $this;
     }
 
-    /**
-     * @param array $aliases
-     * @param bool $replace
-     * @return $this
-     */
-    public function registerAliases(array $aliases, $replace = false)
+    public function registerAliases(array $aliases, bool $replace = false): MiddlewareAliasRegistryInterface
     {
         if ($replace) {
             $this->registry = [];
@@ -51,25 +39,35 @@ class MiddlewareAliasRegistry implements MiddlewareAliasRegistryInterface
 
     /**
      * @param string $alias
-     * @return string|null
+     * @return string
+     * @throws \InvalidArgumentException
      */
-    public function findNameByAlias($alias)
+    public function findNameByAlias(string $alias): string
     {
-        return isset($this->registry[$alias]) ? $this->registry[$alias] : null;
+        if (!$this->aliasExists($alias)) {
+            throw new \InvalidArgumentException("Middleware alias '$alias' has not been registered");
+        }
+
+        return $this->registry[$alias];
     }
 
     /**
      * @param array $middlewares
      * @return array
      */
-    public function resolveAliases(array $middlewares)
+    public function resolveAliases(array $middlewares): array
     {
         return array_map(
             function ($middleware) {
-                return $this->findNameByAlias($middleware) ?: $middleware;
+                return $this->aliasExists($middleware) ? $this->findNameByAlias($middleware) : $middleware;
             },
             $middlewares
         );
+    }
+
+    private function aliasExists(string $alias): bool
+    {
+        return isset($this->registry[$alias]);
     }
 
 }
