@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Upgate\LaravelJsonRpc\ServiceProvider;
 
@@ -6,7 +7,7 @@ use Illuminate\Support\ServiceProvider;
 use Psr\Log\LoggerInterface;
 use Upgate\LaravelJsonRpc\Contract\ServerInterface as JsonRpcServerContract;
 use Upgate\LaravelJsonRpc\Server\RouteDispatcher;
-use Upgate\LaravelJsonRpc\Server\MiddlewareDispatcher;
+use Upgate\LaravelJsonRpc\Server\MiddlewarePipelineDispatcher;
 use Upgate\LaravelJsonRpc\Server\RequestFactory;
 use Upgate\LaravelJsonRpc\Server\Router;
 use Upgate\LaravelJsonRpc\Server\Server;
@@ -23,7 +24,7 @@ class JsonRpcServerServiceProvider extends ServiceProvider
                     new RequestFactory(),
                     new Router(),
                     new RouteDispatcher($this->app),
-                    new MiddlewareDispatcher($this->app),
+                    new MiddlewarePipelineDispatcher($this->app),
                     $this->app->make(LoggerInterface::class)
                 );
             }
